requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.42.2
dj-database-url==2.3.0
Django==5.1.4
django-allauth==65.4.1
django-cloudinary-storage==0.3.0
django-debug-toolbar==5.0.1
django-environ==0.12.0
django-extensions==3.2.3
django-humanize==0.1.2
django-redis==5.4.0
djangorestframework==3.15.2
flake8==7.1.1
gunicorn==23.0.0
humanize==4.12.1
idna==3.10
isort==6.0.0
kombu==5.4.2
mccabe==0.7.0
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
stripe==11.5.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
whitenoise==6.9.0
